Update resource_estimation.py
The resource_estimation.py file provides functions for estimating resources used by quantum circuits. It defines the following functions:

gate_counts: Counts the number of each type of gate used in the quantum circuit.
qubit_count: Counts the number of qubits used in the quantum circuit.
depth: Computes the depth of the quantum circuit.
width: Computes the width of the quantum circuit.
These functions take a QuantumCircuit object as input and return the respective resource estimation. The gate_counts function returns a Counter object, which contains the counts of each gate type.